feat: add content existence check before deletion in DeleteContentUseCase

+++ RediUX-Back/src/modules/content/useCase/deleteContent/DeleteContentUseCase.ts
@@ -1,5 +1,6 @@
 import { inject, injectable } from "tsyringe";
 import { IContentRepository } from "../../repository/interface/IContentRepository";
+import { AppError } from "../../../../shared/erros/AppError";
 
 @injectable()
 class DeleteContentUseCase {
@@ -9,7 +10,13 @@ class DeleteContentUseCase {
   ) {}
 
   async execute(contentId: number): Promise<void> {
-    await this.contentRepository.delete(contentId);
+    const content = await this.contentRepository.findById(contentId);
+
+    if (!content) {
+      throw new AppError("Content not found");
+    }
+
+    return await this.contentRepository.delete(contentId);
   }
 }
